refactor(company): organize imports with index directory file
