+++ src/api/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { DataModule } from '../infrastructure/data/data.module';
-import { CompanyModule } from '../modules/company.module';
+import { CompanyModule } from '../modules';
 
 @Module({
   imports: [DataModule, CompanyModule],
