+++ src/api/controllers/company/index.ts
@@ -1,0 +1,1 @@
+export * from './company.controller';
