+++ src/api/presenters/company/index.ts
@@ -1,0 +1,1 @@
+export * from './read-companies.presenter';
