+++ src/api/transports/company/mappers/index.ts
@@ -1,0 +1,1 @@
+export * from './company-response.mapper';
