+++ src/application/company/outputs/index.ts
@@ -1,0 +1,1 @@
+export * from './read-companies.output';
