+++ src/application/company/outputs/read-companies.output.ts
@@ -1,4 +1,4 @@
-import { Company } from '../../../domain/entities/company';
+import { Company } from '../../../domain/entities';
 
 export class ReadCompanyOutput extends Company {
   id?: string;
