+++ src/application/company/services/index.ts
@@ -1,0 +1,1 @@
+export * from './read-companies.service';
