+++ src/application/company/usecases/index.ts
@@ -1,0 +1,1 @@
+export * from './read-companies-db.interface';
