+++ src/application/company/usecases/read-companies-db.interface.ts
@@ -1,4 +1,4 @@
-import { ReadCompanyOutput } from '../outputs/read-companies.output';
+import { ReadCompanyOutput } from '../outputs';
 
 export class IReadCompaniesDbUseCase {
   readAll: () => Promise<ReadCompanyOutput[]>;
