+++ src/domain/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './base';
+export * from './company';
