+++ src/infrastructure/access/repositories/company/index.ts
@@ -1,0 +1,1 @@
+export * from './company-db.repository';
