+++ src/modules/company.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { CompanyController } from '../api/controllers/company/company.controller';
-import { CompanyPresenter } from '../api/presenters/company/read-companies.presenter';
-import { CompanyResponseMapper } from '../api/transports/company/mappers/company-response.mapper';
-import { ReadCompaniesService } from '../application/company/services/read-companies.service';
-import { CompanyDbRepository } from '../infrastructure/access/repositories/company/company-db.repository';
-import { CompanyModel } from '../infrastructure/access/repositories/company/models/company.model';
+import { CompanyController } from '../api/controllers/company';
+import { CompanyPresenter } from '../api/presenters/company';
+import { CompanyResponseMapper } from '../api/transports/company/mappers';
+import { ReadCompaniesService } from '../application/company/services';
+import { CompanyDbRepository } from '../infrastructure/access/repositories/company';
+import { CompanyModel } from '../infrastructure/access/repositories/company/models';
 
 @Module({
   controllers: [CompanyController],
